Split C codegen helper tests
